Add mokey patch into core
